pyproject.toml
[tool.poetry]
name = "djangorestframework-snatch"
version = "0.1.0"
description = ""
authors = ["bottane4ka <bottane4ka@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
django-rest-framework = "^0.1.0"
psycopg2-binary = "~=2.8"
pydantic = "=1.7.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
